Extract reusable save download helpers
